refactor(worker-javascript): extract Phase 2 helpers from Core.js
Continues the Core.js breakup with six moderate-effort modules. Same
pattern as Phase 1 (composed sibling holding a `core` back-reference,
thin delegating wrapper on Core for every public method/property).
No behavior change.

Modules extracted:
- RendererInstructionBuilder.ts — owns componentsToRender,
  componentsWithChangedChildrenToRender, rendererState; the dast
  instruction stream sent to the renderer
- ProcessQueue.ts — owns processQueue, processing,
  stopProcessingRequests; async request queue and entry-point
  scheduling (executeProcesses, requestAction, requestUpdate,
  requestRecordEvent)
- ComponentLifecycle.ts — stateless: registration, ancestors,
  defining-child splicing, propagation to shadows
- ChildMatcher.ts — child-group matching, adapter substitution,
  rendered-child filtering (recursion guard only)
- DeletionEngine.ts — stateless two-phase component deletion
- ActionTriggerScheduler.ts — owns stateVariableChangeTriggers,
  actionsChangedToActions, originsOfActionsChangedToActions; trigger
  polling and chained-action graph

Core.js drops from 12,909 to 11,253 lines (this PR), 13,837 → 11,253
since the refactor began (~18.7%).
